+++ R/logistftest.R
@@ -26,9 +26,7 @@
 #' @param beta0 Specifies the initial values of the coefficients for the fitting algorithm
 #' @param weights Case weights
 #' @param control Controls parameters for iterative fitting
-#' @param tau  Penalization parameter (if missing: taken from object)
-#' @param col.fit.object Numerical vector containing the positions of the variables to fit, if not
-#' specified: all variables are taken
+#' @param fitcontrol Controls additional parameter for fitting. Default is \code{logistf.fit.control = logistf.fit.control()}
 #' @param ... further arguments passed to logistf.fit
 #'
 #' @return The object returned is of the class logistf and has the following attributes:
@@ -64,7 +62,7 @@
 #' 
 #' 
 logistftest <-
-function(object, test, values, firth = TRUE, beta0, weights, control, tau,  col.fit.object = NULL, ...)
+function(object, test, values, firth = TRUE, beta0, weights, control, fitcontrol, ...)
 {
     call <- match.call()
     formula<-object$formula
@@ -79,7 +77,7 @@ function(object, test, values, firth = TRUE, beta0, weights, control, tau,  col.
     x <- model.matrix(object$formula, model.frame(object))
 
     if (missing(control)) control<-object$control
-    if (missing(tau)) tau<-object$tau
+    if (missing(fitcontrol)) fitcontrol<-object$fitcontrol
     
     cov.name <- labels(x)[[2]]
     if(missing(weights) & !is.null(object$weights)) weight <- object$weights
@@ -99,11 +97,11 @@ function(object, test, values, firth = TRUE, beta0, weights, control, tau,  col.
         coltotest <-1:k
     }
     
-    if(!is.null(col.fit.object)){
-        fit.full<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=col.fit.object, control=control, tau=tau, ...)
+    if(!is.null(fitcontrol$terms.fit)){
+        fit.full<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=fitcontrol$terms.fit, control=control, tau=fitcontrol$tau, ...)
     }
     else {
-        fit.full<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=1:k, control=control, tau=tau, ...)
+        fit.full<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=1:k, control=control, tau=fitcontrol$tau, ...)
     }
     
     if(fit.full$iter>=control$maxit){
@@ -137,7 +135,7 @@ function(object, test, values, firth = TRUE, beta0, weights, control, tau,  col.
         offset1[pos] <- values
     }
     beta <- offset1 
-    fit.null<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=(1:k)[-pos], control=control, init=beta, tau=tau, ...)
+    fit.null<-logistf.fit(x=x, y=y, weight=weight, offset=offset, firth, col.fit=(1:k)[-pos], control=control, init=beta, tau=fitcontrol$tau, ...)
 
     if(fit.null$iter>=control$maxit){
         warning(paste("logistftest: Maximum number of iterations for null model exceeded. Try to increase the number of iterations by passing 'logistf.control(maxit=...)' to parameter control"))
@@ -148,7 +146,7 @@ function(object, test, values, firth = TRUE, beta0, weights, control, tau,  col.
     offset1[ - pos] <- NA
     names(offset1) <- cov.name
     fit <- list(testcov = offset1, loglik = loglik, df = k2, prob = 1 - pchisq(2 *
-        diff(loglik), k2), call = match.call(), beta = beta, tau = tau)
+        diff(loglik), k2), call = match.call(), beta = beta)
     if(firth) {
         fit$method <- "Penalized ML"
     }