+++ R/summary.flac.R
@@ -19,8 +19,8 @@ summary.flac <- function(object, ...)
   
   #consider for wald only covariance matrix with columns corresponding to variables in terms.fit
   call <- object$call
-  if(!is.null(call$terms.fit) | !is.null(terms.fit)){
-    terms.fit <- c(call$terms.fit,terms.fit)[!is.null(c(call$terms.fit,terms.fit))]
+  if(!is.null(object$fitcontrol$terms.fit)){
+    terms.fit <- object$fitcontrol$terms.fit[!is.null(object$fitcontrol$terms.fit)]
     terms.fit <- eval(terms.fit, parent.frame())
     loc <- match(terms.fit, object$terms)
     var.red <- object$var[loc,loc]